finished blue requirements
