+++ code/src/components/ThoughtForm.js
@@ -2,16 +2,30 @@ import React from 'react';
 
 const ThoughtForm = ({ onFormSubmit, newThought, setNewThought }) => {
   return (
-    <form onSubmit={onFormSubmit}>
-      <label htmlFor='newThought'>Type your thought </label>
+    <form
+      onSubmit={onFormSubmit}
+      className='text-container'
+      id='form-container'
+    >
+      <label htmlFor='newThought'>What´s making you happy right now? </label>
       <input
         id='newThought'
         type='text'
         value={newThought}
         onChange={(e) => setNewThought(e.target.value)}
       />
-      <button disabled={newThought.length < 5} type='submit'>
-        Send thought
+      <button
+        className='send-thought'
+        disabled={newThought.length < 5}
+        type='submit'
+      >
+        <span role='img' aria-label='heart'>
+          ❤️&ensp;
+        </span>
+        Send happy thought
+        <span role='img' aria-label='heart'>
+          &ensp;❤️
+        </span>
       </button>
     </form>
   );
