+++ code/src/components/ThoughtItem.js
@@ -3,15 +3,23 @@ import moment from 'moment';
 
 const ThoughtItem = ({ thought, onLikesIncrease }) => {
   return (
-    <div>
+    <div className='text-container'>
       <p>{thought.message}</p>
-      <button onClick={() => onLikesIncrease(thought._id)}>
-        {' '}
-        &hearts; {thought.hearts}
-      </button>
-      <p className='date'>
-        - Created at: {moment(thought.createdAt).fromNow()}
-      </p>
+      <div className='thought-data'>
+        <span className='likes-counter'>
+          <button
+            className='likes-button'
+            onClick={() => onLikesIncrease(thought._id)}
+          >
+            {' '}
+            <span className='button-heart' role='img' aria-label='heart'>
+              ❤️&ensp;
+            </span>{' '}
+          </button>
+          <p>x {thought.hearts}</p>
+        </span>
+        <p className='date'>{moment(thought.createdAt).fromNow()}</p>
+      </div>
     </div>
   );
 };
