+++ code/src/index.css
@@ -1,10 +1,11 @@
 body {
   margin: 10px;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, 'Inconsolata', 'Segoe UI',
+    'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans',
+    'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  font-size: 18px;
 }
 
 code {
@@ -12,9 +13,97 @@ code {
     monospace;
 }
 
-.date {
-  font-style: italic;
+:root {
+  --base: #fff;
+  --primary: #f2f2f2;
+  --secondary: #f2a7a7;
 }
+
+h1 {
+  text-align: center;
+}
+
+.text-container {
+  border: 2px solid black;
+  box-shadow: 6px 6px 1px 1px #000;
+  display: inline-flex;
+  flex-direction: column;
+  font-weight: 600;
+  margin: 5%;
+  padding: 5%;
+  width: 80%;
+}
+
+#form-container {
+  background: var(--primary);
+}
+
+/* ----------------------------------------------- */
+/*                      FORM			           */
+/* ----------------------------------------------- */
+
+input {
+  font-family: 'Inconsolata';
+  font-size: 16px;
+  font-weight: 500px;
+  height: 8vh;
+  margin-top: 2vh;
+  margin-bottom: 2vh;
+}
+
+.send-thought {
+  background: var(--secondary);
+  border: none;
+  border-radius: 12px;
+  color: #000;
+  -webkit-border-radius: 12px;
+  -moz-border-radius: 12px;
+  -ms-border-radius: 12px;
+  -o-border-radius: 12px;
+  font-family: 'Inconsolata';
+  font-weight: 600;
+  text-align: left;
+  padding: 3%;
+  width: 60%;
+}
+
+/* ----------------------------------------------- */
+/*                     THOUGHTS			           */
+/* ----------------------------------------------- */
+
+.likes-counter {
+  display: inline-flex;
+  column-gap: 10px;
+}
+
+.likes-button {
+  border: none;
+  border-radius: 50%;
+  -webkit-border-radius: 50%;
+  -moz-border-radius: 50%;
+  -ms-border-radius: 50%;
+  -o-border-radius: 50%;
+  height: 40px;
+  margin-top: 3px;
+  width: 40px;
+}
+
+.button-heart {
+  padding-left: 18%;
+}
+
+.thought-data {
+  display: inline-flex;
+  justify-content: space-between;
+}
+
+.date {
+  color: rgb(167, 160, 160);
+}
+
+/* ----------------------------------------------- */
+/*                   LOADING OVERLAY	           */
+/* ----------------------------------------------- */
 
 .loading-overlay {
   position: absolute;
@@ -22,7 +111,7 @@ code {
   right: 0;
   bottom: 0;
   left: 0;
-  background: #fff;
+  background: var(--base);
   display: flex;
   justify-content: center;
   align-items: center;
@@ -32,7 +121,7 @@ code {
   width: 50px;
   height: 50px;
   border: 5px solid #dcdcdc;
-  border-left: 5px solid #fff;
+  border-left: 5px solid var(--base);
   border-radius: 50%;
   -webkit-border-radius: 50%;
   -moz-border-radius: 50%;
@@ -63,5 +152,46 @@ code {
     -moz-transform: rotate(360deg);
     -ms-transform: rotate(360deg);
     -o-transform: rotate(360deg);
+  }
+}
+
+/* ----------------------------------------------- */
+/*               MEDIA QUERIES                     */
+/* ----------------------------------------------- */
+
+@media only screen and (min-width: 768px) {
+  body {
+    font-size: 25px;
+    margin-left: auto;
+    margin-right: auto;
+    width: 50%;
+  }
+}
+
+@media only screen and (min-width: 1024px) {
+  h1 {
+    margin-top: 5%;
+    text-align: center;
+  }
+
+  .text-container {
+    margin-bottom: 0;
+    margin-top: 3%;
+    padding: 3%;
+  }
+
+  .send-thought {
+    border-radius: 25px;
+    font-size: 20px;
+    width: 40%;
+    -webkit-border-radius: 25px;
+    -moz-border-radius: 25px;
+    -ms-border-radius: 25px;
+    -o-border-radius: 25px;
+  }
+
+  .likes-button {
+    height: 60px;
+    width: 55px;
   }
 }
